Add readme and basic tile images to use in the renderer
Apparently grafx2 segfaults when saving files as PNGs, so I'm reduced to using
Pixilart, which is annoying because it's a free website with nonstop ads in the
bottom right corner and right-click doesn't work properly (erases to transparent
instead of using the background color like the tooltip says. However, it does
work, so I can't complain too much.